Create express server and static and S3 endpoints.
This creates an express server which is compatible with personal-server.

The express server has two endpoints: One for creating the light markup,
and a second for supplying API data which will populate the static pages

The API data is formed from periodic requests to digitalocean for
information about our data, which is then sorted for the user to
retrieve at the /data endpoint.