Style Lectures to be uniform in size. Solves #4
While we could use a conventional table which adapts column size
depending on the largest contents, that would mean that adding a new
course to the queue can radically change the look of the table in one
go. To make sizing more reliable, all columns are statically sized
with the understanding that course IDs will not exceed 5 chars.